requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
numpy==1.23.4
openai==0.25.0
opencv-python==4.6.0.66
Pillow==9.3.0
pyrsistent==0.19.2
pytesseract==0.3.10
requests==2.28.1
streamlit==1.14.0
